# Optimize HttpDateParser.TryStringToDate