--- a/src/libraries/System.Net.Http/src/System/Net/Http/Headers/ContentDispositionHeaderValue.cs
+++ b/src/libraries/System.Net.Http/src/System/Net/Http/Headers/ContentDispositionHeaderValue.cs
@@ -322,7 +322,7 @@ namespace System.Net.Http.Headers
                 {
                     dateString = dateString.Slice(1, dateString.Length - 2);
                 }
-                if (HttpDateParser.TryStringToDate(dateString, out date))
+                if (HttpDateParser.TryParse(dateString, out date))
                 {
                     return date;
                 }